@@ -1,5 +1,5 @@
 default_app_config = 'aadiscordbot.apps.AADiscordBotConfig'
 
-__version__ = '0.5.3'
+__version__ = '3.0.0'
 __title__ = "AA Discordbot"
-__branch__ = 'Stable'
+__branch__ = 'slash-branch'
