@@ -14,6 +14,7 @@ from . import bot_tasks
 from aadiscordbot.app_settings import DISCORD_BOT_ACCESS_DENIED_REACT, DISCORD_BOT_PREFIX
 
 import discord
+from discord import TextChannel, Role, CategoryChannel
 from discord.ext import commands, tasks
 
 import django
@@ -25,13 +26,14 @@ from kombu import Connection, Queue, Consumer
 from socket import timeout
 import concurrent.futures
 
+
 description = """
 AuthBot is watching...
 """
 
 logger = logging.getLogger(__name__)
 
-queuename="aadiscordbot"
+queuename = "aadiscordbot"
 queue_keys = [f"{queuename}",
               f"{queuename}\x06\x161",
               f"{queuename}\x06\x162",
@@ -42,6 +44,7 @@ queue_keys = [f"{queuename}",
               f"{queuename}\x06\x167",
               f"{queuename}\x06\x168",
               f"{queuename}\x06\x169"]
+
 
 class AuthBot(commands.Bot):
     def __init__(self):
@@ -57,17 +60,20 @@ class AuthBot(commands.Bot):
         )
 
         self.redis = None
-        self.redis = self.loop.run_until_complete(aioredis.create_pool(getattr(settings, "BROKER_URL", "redis://localhost:6379/0"), minsize=5, maxsize=10))
+        self.redis = self.loop.run_until_complete(aioredis.create_pool(getattr(
+            settings, "BROKER_URL", "redis://localhost:6379/0"), minsize=5, maxsize=10))
         print('redis pool started', self.redis)
         self.client_id = client_id
         self.session = aiohttp.ClientSession(loop=self.loop)
         self.tasks = []
 
-        self.message_connection = Connection(getattr(settings, "BROKER_URL", 'redis://localhost:6379/0'))
+        self.message_connection = Connection(
+            getattr(settings, "BROKER_URL", 'redis://localhost:6379/0'))
         queues = []
         for que in queue_keys:
             queues.append(Queue(que))
-        self.message_consumer = Consumer(self.message_connection, queues, callbacks=[self.on_queue_message], accept=['json'])
+        self.message_consumer = Consumer(self.message_connection, queues, callbacks=[
+                                         self.on_queue_message], accept=['json'])
 
         django.setup()
 
@@ -79,17 +85,18 @@ class AuthBot(commands.Bot):
                     print(f"Failed to load cog {cog}", file=sys.stderr)
                     traceback.print_exc()
 
-
     def on_queue_message(self, body, message):
         print('RECEIVED MESSAGE: {0!r}'.format(body))
         try:
             task_headers = message.headers
+            _args = body[0]
+            _kwargs = body[1]
 
             if 'aadiscordbot.tasks.' in task_headers["task"]:
                 task = task_headers["task"].replace("aadiscordbot.tasks.", '')
                 task_function = getattr(bot_tasks, task, False)
                 if task_function:
-                    self.tasks.append((task_function, body[0]))
+                    self.tasks.append((task_function, _args, _kwargs))
                     if not bot_tasks.run_tasks.is_running():
                         bot_tasks.run_tasks.start(self)
                 else:
@@ -111,12 +118,11 @@ class AuthBot(commands.Bot):
                                     type=discord.ActivityType.watching,
                                     state="Monitoring",
                                     details="Waiting for Shenanigans!",
-                                    emoji={"name":":smiling_imp:"}
+                                    emoji={"name": ":smiling_imp:"}
                                     )
         await self.change_presence(activity=activity)
 
         self.poll_queue.start()
-
         logger.info("Ready")
 
     async def process_commands(self, message):
@@ -141,29 +147,8 @@ class AuthBot(commands.Bot):
                 with self.message_consumer:
                     self.message_connection.drain_events(timeout=0.01)
             except timeout as e:
-                #logging.exception(e)
+                # logging.exception(e)
                 message_avail = False
-
-    async def queue_consumer(self, task):
-        logger.debug("Queue Consumer has started")
-        try:
-            task_headers = task["headers"]
-            task_header_args = eval(task_headers["argsrepr"])
-
-            if 'aadiscordbot.tasks.' in task_headers["task"]:
-                task = task_headers["task"].replace("aadiscordbot.tasks.", '')
-                task_function = getattr(bot_tasks, task, False)
-                if task_function:
-                    await task_function(self, task_header_args)
-                else:
-                    logger.debug("No bot_task for that auth_task?")
-            else:
-                logger.debug("i got an invalid auth_task")
-
-        except Exception as e:
-            logger.error("Queue Consumer Failed")
-            logger.error(e, exc_info=1)
-
 
     async def on_resumed(self):
         print("Resumed...")
@@ -183,7 +168,8 @@ class AuthBot(commands.Bot):
             return await ctx.send(error)
         elif isinstance(error, commands.BotMissingPermissions):
             await ctx.send(
-                "Sorry, I don't have the required permissions to do that here:\n{0}".format(error.missing_perms)
+                "Sorry, I don't have the required permissions to do that here:\n{0}".format(
+                    error.missing_perms)
             )
         elif isinstance(error, commands.MissingPermissions):
             await ctx.message.add_reaction(chr(DISCORD_BOT_ACCESS_DENIED_REACT))
@@ -195,22 +181,5 @@ class AuthBot(commands.Bot):
             await ctx.message.add_reaction(chr(0x274C))
 
     def run(self):
+        # self.load_extension("aadiscordbot.slash.admin")
         super().run(settings.DISCORD_BOT_TOKEN, reconnect=True)
-
-## Fetching Tasks from celery queue for the message sending loop
-async def get_task(bot):
-    logger.debug("im getting a task")
-    try:
-        task = await bot.redis.execute("brpop", queuename, *queue_keys, 1)
-        if task != None:
-            logger.info('ive got a task')
-            logger.debug(task)
-            return json.loads(task[1])
-        else:
-            logger.debug("No tasks in queue")
-            return False
-
-    except Exception as e:
-        logger.error("Get Task Failed")
-        logger.error(e)
-        pass
