@@ -35,8 +35,8 @@ class Timers(commands.Cog):
         next_timer = Timer.objects.filter(corp_timer=False,
                                           eve_time__gte=datetime.datetime.utcnow().replace(tzinfo=timezone.utc)).first()
         time_until = pendulum.now(tz="UTC").diff_for_humans(
-                            next_timer.eve_time, absolute=True
-                        )
+            next_timer.eve_time, absolute=True
+        )
         embed = Embed(title="Next Timer")
         embed.description = next_timer.details
         if next_timer.objective == "Friendly":
@@ -46,7 +46,8 @@ class Timers(commands.Cog):
         else:
             embed.colour = Color.white()
         try:
-            embed.set_footer(text="Added By {0}".format(next_timer.eve_character.character_name))
+            embed.set_footer(text="Added By {0}".format(
+                next_timer.eve_character.character_name))
         except Exception as e:
             logger.error(e)
             pass
