@@ -1,7 +1,7 @@
+from aadiscordbot import launcher
+import asyncio
 import os
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "myauth.settings.local")
-import asyncio
-from aadiscordbot import launcher
 
 
 def main():
